Add View class and templates/ directory
The View class allows you to create a "View" from a specific template file and
bind variables to the page. This could be helpful for anytime we want to alter
some HTML before we send it to the client. I left a sample template called
base.phtml in the templates directory.