Add Reddit class for Reddit API operations.
This class servers as a wrapper for the Reddit API. At the moment, it only has
methods for getting comments and submissions (posts) from an array of
usernames. Some other functionality this could hold includes sending an
authentication request for a reddit user, posting a comment reply on behalf of
a user, "saving" a post on behalf of a user, etc.