Fix routing for static files
When our page was trying to grab our own css and js files, the request would be
picked up by our PHP controller and the originally requested file would not be
returned.

This adds a check in `app.js` that makes sure the `REQUEST_URI` does not
contain 'public'. If it does, it bypasses our PHP routing and serves the static
file that is requested from `public/`.

If you add any new css or js files, please place them in either
`public/js/yourfile.js` or `public/css/yourfile.css`.