Identify with userids instead of usernames, use sessions for identification.
This is a pretty involved commit, but it accomplishes a combination of two
pretty big refactorings.

1. Identifying users with userids instead of usernames.

This adds a new userid column to the users table which is also the new primary
key. All functionality that required a username to determine which user to do a
certains operation on now require a userid. Usernames also have a new unique
contraint so that you can't create multiple users with the same name.

2. Using sessions cookies to track who has logged in.

There is a new sessions table that has sessionid, userid, and expiration
columns. Instead of the old "user=username" cookie that was used to identify
users (*shudder*), users are given a session cookie that corresponds with a row
in the sessions table. Users should only have to log in once a month. There
should be a way to disable sessions on logout, but that's something for
sessions v.2.