Reload on settings change
If the user tries to go back to the following or subreddit tabs after a
settings change, the app will reload itself to reload the new changes. A more
polished way to do this in the future would be to show the loading screen while
pulling in more changes through an AJAX request.