Reload outdated content on tab switch
This adds a status object for different types of settings and whether or not
they are "dirty". Settings groups are on a view basis, so they consist of
following, subreddits, and settings themselves. If you change one or more of
them, it marks them as dirty.

When the user clicks on a different view tab after changing settings, a loading
image is thrown up and the out of date sections are refreshed.

This is awesome because it means that the user doesn't have to refresh the page
to see their new settings in effect.